Move InMemoryAppPreferences
